listas las rutas de crear post
