+++ app.js
@@ -1,14 +1,14 @@
 const express = require("express");
 const morgan = require("morgan");
 const cors = require("cors");
-
 const bcrypt = require("bcryptjs");
 const createToken = require("jsonwebtoken");
 require("dotenv").config();
 const indexUserModel = require("./src/db.js");
 const server = express();
-const routes = require("./src/routes/index.js");
+const routes = require("./src/routes/index");
 const authenticationToken = require("./src/controllers/token-jimi");
+const { default: axios } = require("axios");
 
 server.use(express.urlencoded({ extended: true }));
 const auth = express();
@@ -72,15 +72,16 @@ auth.use(function (req, res, next) {
   }
 });
 
-//ruta para ignorar middlewares
+//ruta para ignorar middleware
 //server.use("/", routes);
 
 //ruta para probar token jwt
-server.use("/", auth, routes);
+// server.use("/", auth, routes);
 
 //ruta para probar refresh api token
-// server.use("/", authenticationToken, routes);
+server.use("/", authenticationToken, routes);
 
 //ruta ambos middlewares
-// server.use("/", [auth,authenticationToken], routes);
+// server.use("/", [auth, authenticationToken], routes);
+
 module.exports = server;
