--- configToken/config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  llave: "HenryB@ck_6",
-};
