+++ src/routes/Api-Jimi/coords-of-day.js
@@ -2,8 +2,8 @@ const { default: axios } = require("axios");
 const { Router } = require("express");
 const Md5 = require("md5");
 const utf8 = require("utf8");
-const { APP_KEY, APP_SECRET, USER_ID } = process.env;
-const { Token } = require("../db");
+const { APP_KEY, APP_SECRET } = process.env;
+const { Token } = require("../../db");
 const router = Router();
 
 router.post("/", async function (req, res) {
@@ -14,17 +14,15 @@ router.post("/", async function (req, res) {
     app_key: APP_KEY,
     timestamp: new Date().toISOString().slice(0, 19).replace("T", " "),
     format: "json",
-    method: "jimi.device.media.URL",
+    method: "jimi.device.track.list",
     v: "1.0",
     sign_method: "md5",
     access_token: tokenPassword.token,
     imei: "862798050059324",
-    media_type: "2",
-    camera: "2",
-    // start_time: "2021-07-20 18:43:26",
-    // end_time: "2021-07-20 18:53:26",
+    map_type: "GOOGLE",
+    begin_time: "2021-12-13 18:12:00",
+    end_time: "2021-12-13 18:13:00",
   };
-
   //str de parametros ordenados alfabeticamente y unidos
   let temp = utf8.encode(
     Object.entries(paramsSing).sort().join().replace(/,/g, "")
@@ -32,24 +30,20 @@ router.post("/", async function (req, res) {
   //genero el sign concatenando los datos, "hasheado" con md5 y modificando todo eso a mayusculas
   let app_secret = APP_SECRET;
   const sign = Md5(app_secret + temp + app_secret).toUpperCase();
-
   // creo la query de parametros de la peticion
   let urlencoded = new URLSearchParams();
   urlencoded.append("sign", sign);
   urlencoded.append("app_key", paramsSing.app_key);
+  urlencoded.append("timestamp", paramsSing.timestamp);
   urlencoded.append("format", paramsSing.format);
-  urlencoded.append("v", paramsSing.v);
   urlencoded.append("method", paramsSing.method);
+  urlencoded.append("v", paramsSing.v);
   urlencoded.append("sign_method", paramsSing.sign_method);
   urlencoded.append("access_token", paramsSing.access_token);
-  urlencoded.append("timestamp", paramsSing.timestamp);
   urlencoded.append("imei", paramsSing.imei);
-  urlencoded.append("media_type", paramsSing.media_type);
-  urlencoded.append("camera", paramsSing.camera);
-  // urlencoded.append("start_time", paramsSing.start_time);
-  // urlencoded.append("end_time", paramsSing.end_time);
-
-  // objeto que define las propiedades de la peticion
+  urlencoded.append("end_time", paramsSing.end_time);
+  urlencoded.append("begin_time", paramsSing.begin_time);
+  urlencoded.append("map_type", paramsSing.map_type);
   var requestOptions = {
     method: "POST",
     Accept: "application/json",
@@ -60,6 +54,7 @@ router.post("/", async function (req, res) {
     },
     params: urlencoded,
   };
+  console.log(requestOptions.params);
 
   //hago la peticion y devuelvo la info a postman
   axios("http://open.10000track.com/route/rest", requestOptions)
